api/requirements.core.txt
Django==4.2.5
django-ninja==0.22.2
knight-auth==0.2.0
psycopg2-binary==2.9.7

